requirements.txt
pandas==1.3.3
pyarrow==8.0.0
snowflake-connector-python==2.8.3
snowflake-sqlalchemy==1.4.4
streamlit==1.15.2